backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.8
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.12
django-cors-headers==3.10.1
django-import-export==2.7.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
persiantools==3.0.0
psycopg2==2.9.3
PyJWT==1.7.1
pyparsing==3.0.7
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tablib==3.1.0
torch==1.10.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
